requirements.txt
streamlit
streamlit_extras
folium
streamlit_dynamic_filters
streamlit_folium
plotly
geopandas
openpyxl
sodapy
Unidecode
leila